LocalRAW/requirements.txt
PySide6
rawpy
numpy
opencv-python
Pillow--